Report History page implemented
