--- frontend/app/src/main/java/com/example/tweetseek/identification/database/ReportInfo.kt
@@ -1,4 +1,0 @@
-package com.example.tweetseek.identification.database
-
-class ReportInfo {
-}
