+++ frontend/app/src/main/java/com/example/tweetseek/models/BirdReport.kt
@@ -1,0 +1,3 @@
+package com.example.tweetseek.models
+
+data class BirdReport(val imageUrl: String, val birdName: String)
