+++ frontend/app/src/main/res/layout/home_page.xml
@@ -10,7 +10,7 @@
     android:layout_height="match_parent"
     android:orientation="vertical"
     android:weightSum="10"
-    tools:context=".HomePage">
+    tools:context=".activity.HomePageActivity">
 
     <ImageView
         android:id="@+id/home_logo"
