+++ frontend/app/src/main/res/layout/input_management.xml
@@ -7,7 +7,7 @@
     android:layout_height="match_parent"
     android:padding="24dp"
     android:background="@android:color/white"
-    tools:context=".MainActivity">
+    tools:context=".activity.MainActivity">
 
     <!-- Back Button -->
     <com.google.android.material.button.MaterialButton
