+++ frontend/app/src/main/res/layout/login_page.xml
@@ -10,7 +10,7 @@
 
     android:orientation="vertical"
     android:gravity="center"
-    tools:context=".LoginActivity">
+    tools:context=".activity.LoginActivity">
     <!-- Logo Input -->
     <ImageView
         android:id="@+id/login_logo"
