+++ frontend/app/src/main/res/layout/result_page.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@drawable/cloudy_bg"
-    tools:context=".RegisterActivity">
+    tools:context=".activity.RegisterActivity">
 
     <ImageButton
         android:id="@+id/exitButton"
