+++ frontend/app/src/main/res/values/strings.xml
@@ -27,4 +27,7 @@
     <string name="new_password">new password</string>
     <string name="email_example_com">email@example.com</string>
     <string name="create_an_account">create an account</string>
+    <string name="identification_history">Identification History</string>
+    <string name="bird_name">Bird Name</string>
+    <string name="bird_image" />
 </resources>
